fix：增加s3的桶信息和代码解耦
